Milestone 1 BDF Almost Complete
I can't figure out how to use the 3 input muxes and I don't want to break something. So everything else is done except the two muxes on the inputs to the ALU. The signals are ready, all I need someone to do is put in the muxes themselves and connect the things to them.