Added PC Functionality, and Cycle Counter
I added the various muxes and adders for the PC+4+Branch line.

Muxes:
-BranchDetection signal is tied low on the Branch vs. PC+4 mux.
- ZeroSignExtend signal is tied low on the SignExt vs. ZeroSignExt mux.

Added ThirtyTwoBitAdder.v
- Adds ID_PCplus4 and Shifted, Sign Extended, Immediate number.
-Output is called BranchedPC
-Drives "high" side of Branch vs. PC+4 mux.

Cycle Counter:
-Used a PC module and a new module called CycleIncrement
-CycleIncrement just adds 1 to the input, it is clocked.
-Lower 16 bits of CycleCounter goes to VGA block.

Possible Bugs:
CycleIncrement is clocked. Does it need to be?
The PC+4 going driving the "low side" of Branch vs. PC+4 mux is NOT the pipelined version of PC+4, does it need to be?